feat: refactoring server fn and lib
