+++ app/routes/_protected/_compte/compte/entreprises/index.tsx
@@ -1,9 +1,5 @@
-import { queryOptions, useMutation, useSuspenseQuery } from "@tanstack/react-query";
-import { createFileRoute, Link, redirect, useRouter } from "@tanstack/react-router";
-import { createServerFn, useServerFn } from "@tanstack/react-start";
-import { getWebRequest } from "@tanstack/react-start/server";
-import { unknown } from "better-auth";
-import { eq, inArray } from "drizzle-orm";
+import { useMutation, useSuspenseQuery } from "@tanstack/react-query";
+import { createFileRoute, Link } from "@tanstack/react-router";
 import { useState } from "react";
 import { toast } from "sonner";
 import { CopyIcon } from "~/components/icons/copy";
@@ -15,67 +11,8 @@ import {
 	DialogTitle,
 	DialogTrigger,
 } from "~/components/ui/dialog";
-import { db } from "~/db";
-import { categoriesTable } from "~/db/schema/categories";
-import { companiesTable } from "~/db/schema/companies";
-import { companyCategoriesTable } from "~/db/schema/company-categories";
-import { auth } from "~/lib/auth";
-
-export const getUserCompanies = createServerFn({ method: "GET" }).handler(async () => {
-	const request = getWebRequest();
-	if (!request) return;
-
-	const session = await auth.api.getSession({ headers: request.headers });
-	if (!session) throw redirect({ to: "/login" });
-
-	// Get companies
-	const companiesQuery = await db
-		.select({
-			id: companiesTable.id,
-			name: companiesTable.name,
-			description: companiesTable.description,
-			status: companiesTable.status,
-			siret: companiesTable.siret,
-		})
-		.from(companiesTable)
-		.where(eq(companiesTable.user_id, session.user.id));
-
-	// Then get categories in a separate query
-	const categories = await db
-		.select({
-			company_id: companyCategoriesTable.company_id,
-			category_id: categoriesTable.id,
-			category_name: categoriesTable.name,
-		})
-		.from(companyCategoriesTable)
-		.leftJoin(categoriesTable, eq(categoriesTable.id, companyCategoriesTable.category_id))
-		.where(
-			inArray(
-				companyCategoriesTable.company_id,
-				companiesQuery.map((c) => c.id),
-			),
-		);
-
-	return companiesQuery.map((company) => ({
-		...company,
-		categories: categories.filter((c) => c.company_id === company.id),
-	}));
-});
-
-export const deleteCompany = createServerFn({ method: "POST" })
-	.validator((data: string) => data as string)
-	.handler(async ({ data }) => {
-		try {
-			await db.delete(companiesTable).where(eq(companiesTable.id, data));
-		} catch (error) {
-			console.error(error);
-		}
-	});
-
-const userCompaniesQueryOptions = queryOptions({
-	queryKey: ["user", "companies"],
-	queryFn: () => getUserCompanies(),
-});
+import { deleteCompany } from "~/lib/api/companies";
+import { userCompaniesQueryOptions } from "~/lib/api/user";
 
 export const Route = createFileRoute("/_protected/_compte/compte/entreprises/")({
 	loader: async ({ context }) => {
