+++ app/routes/api/auth/$.ts
@@ -1,5 +1,5 @@
 import { createAPIFileRoute } from "@tanstack/react-start/api";
-import { auth } from "~/lib/auth";
+import { auth } from "~/lib/auth/auth.server";
 
 export const APIRoute = createAPIFileRoute("/api/auth/$")({
 	GET: ({ request }) => {
